Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	policy/src/com/android/internal/policy/impl/PhoneWindowManager.java
	services/java/com/android/server/wm/WindowAnimator.java

+++ packages/SystemUI/src/com/android/systemui/statusbar/policy/ExtensibleKeyButtonView.java
@@ -8,6 +8,7 @@ import android.content.ActivityNotFoundException;
 import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.ResolveInfo;
 import android.content.ServiceConnection;
 import android.hardware.input.InputManager;
 import android.os.Handler;
@@ -24,9 +25,10 @@ import android.view.KeyCharacterMap;
 import android.view.KeyEvent;
 import android.view.View;
 import android.view.ViewConfiguration;
+import android.widget.Toast;
 
 import com.android.internal.statusbar.IStatusBarService;
-import com.android.systemui.statusbar.phone.NavigationBarView;
+import com.android.systemui.statusbar.WidgetView;
 import com.android.systemui.statusbar.policy.KeyButtonView;
 import com.android.systemui.R;
 
@@ -134,8 +136,18 @@ public class ExtensibleKeyButtonView extends KeyButtonView {
 
     Runnable mKillTask = new Runnable() {
         public void run() {
+            final Intent intent = new Intent(Intent.ACTION_MAIN);
+            String defaultHomePackage = "com.android.launcher";
+            intent.addCategory(Intent.CATEGORY_HOME);
+            final ResolveInfo res = mContext.getPackageManager().resolveActivity(intent, 0);
+            if (res.activityInfo != null && !res.activityInfo.packageName.equals("android")) {
+                defaultHomePackage = res.activityInfo.packageName;
+            }
             String packageName = mActivityManager.getRunningTasks(1).get(0).topActivity.getPackageName();
-            mActivityManager.forceStopPackage(packageName);
+            if (!defaultHomePackage.equals(packageName)) {
+                    mActivityManager.forceStopPackage(packageName);
+                    Toast.makeText(mContext, R.string.app_killed_message, Toast.LENGTH_SHORT).show();
+            }
         }
     };
 
@@ -175,18 +187,18 @@ public class ExtensibleKeyButtonView extends KeyButtonView {
                 takeScreenshot();
                 return;
 
-        	} else if (mClickAction.equals(ACTION_KILL)) {
+                } else if (mClickAction.equals(ACTION_KILL)) {
 
-        		mHandler.postDelayed(mKillTask,ViewConfiguration.getGlobalActionKeyTimeout());
-        		return;
+                        mHandler.postDelayed(mKillTask,ViewConfiguration.getGlobalActionKeyTimeout());
+                        return;
 
-        	} else if (mClickAction.equals(ACTION_WIDGETS)) {
-        		Intent toggleWidgets = new Intent(
-                        NavigationBarView.WidgetReceiver.ACTION_TOGGLE_WIDGETS);
+                } else if (mClickAction.equals(ACTION_WIDGETS)) {
+                        Intent toggleWidgets = new Intent(
+                        WidgetView.WidgetReceiver.ACTION_TOGGLE_WIDGETS);
                 mContext.sendBroadcast(toggleWidgets);
                 return;
-        	} else {  // we must have a custom uri
-        		 try {
+                } else {  // we must have a custom uri
+                         try {
                      Intent intent = Intent.parseUri(mClickAction, 0);
                      intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
                      getContext().startActivity(intent);
@@ -199,52 +211,52 @@ public class ExtensibleKeyButtonView extends KeyButtonView {
             return;
         }
     };
-    
+
     private OnLongClickListener mLongPressListener = new OnLongClickListener() {
 
         @Override
         public boolean onLongClick(View v) {
-        	if (mLongpress == null) {
-        		return true;
-        	}
-        	if (mLongpress.equals(ACTION_NULL)) {
-        		// attempt to keep long press functionality of 'keys' if
-        		// they haven't been overridden.
+                if (mLongpress == null) {
+                        return true;
+                }
+                if (mLongpress.equals(ACTION_NULL)) {
+                        // attempt to keep long press functionality of 'keys' if
+                        // they haven't been overridden.
                 return true;
-        	} else if (mLongpress.equals(ACTION_HOME)) {
-        		injectKeyDelayed(KeyEvent.KEYCODE_HOME);
-        		return true;
-        	} else if (mLongpress.equals(ACTION_BACK)) {
-        		injectKeyDelayed(KeyEvent.KEYCODE_BACK);
-        		return true;
-        	} else if (mLongpress.equals(ACTION_SEARCH)) {
-        		injectKeyDelayed(KeyEvent.KEYCODE_SEARCH);
-        		return true;
-        	} else if (mLongpress.equals(ACTION_MENU)) {
-        		injectKeyDelayed(KeyEvent.KEYCODE_MENU);
-        		return true;
-        	} else if (mLongpress.equals(ACTION_POWER)) {
-        		injectKeyDelayed(KeyEvent.KEYCODE_POWER);
-        		return true;
+                } else if (mLongpress.equals(ACTION_HOME)) {
+                        injectKeyDelayed(KeyEvent.KEYCODE_HOME);
+                        return true;
+                } else if (mLongpress.equals(ACTION_BACK)) {
+                        injectKeyDelayed(KeyEvent.KEYCODE_BACK);
+                        return true;
+                } else if (mLongpress.equals(ACTION_SEARCH)) {
+                        injectKeyDelayed(KeyEvent.KEYCODE_SEARCH);
+                        return true;
+                } else if (mLongpress.equals(ACTION_MENU)) {
+                        injectKeyDelayed(KeyEvent.KEYCODE_MENU);
+                        return true;
+                } else if (mLongpress.equals(ACTION_POWER)) {
+                        injectKeyDelayed(KeyEvent.KEYCODE_POWER);
+                        return true;
                 } else if (mLongpress.equals(ACTION_IME)) {
                         getContext().sendBroadcast(new Intent("android.settings.SHOW_INPUT_METHOD_PICKER"));
                         return true;
             } else if (mLongpress.equals(ACTION_SCREENSHOT)) {
                 takeScreenshot();
                 return true;
-        	} else if (mLongpress.equals(ACTION_KILL)) {
-        		mHandler.post(mKillTask);
-        		return true;
+                } else if (mLongpress.equals(ACTION_KILL)) {
+                        mHandler.post(mKillTask);
+                        return true;
             } else if (mLongpress.equals(ACTION_WIDGETS)) {
                 Intent toggleWidgets = new Intent(
-                        NavigationBarView.WidgetReceiver.ACTION_TOGGLE_WIDGETS);
+                        WidgetView.WidgetReceiver.ACTION_TOGGLE_WIDGETS);
                 mContext.sendBroadcast(toggleWidgets);
                 return true;
-        	} else if (mLongpress.equals(ACTION_RECENTS)) {
-        		try {
+                } else if (mLongpress.equals(ACTION_RECENTS)) {
+                        try {
                     mBarService.toggleRecentApps();
-                } catch (RemoteException e) {   
-                	// let it go.
+                } catch (RemoteException e) {
+                        // let it go.
                 }
                 return true;
             } else if (mClickAction.equals(ACTION_NOTIFICATIONS)) {
@@ -255,18 +267,18 @@ public class ExtensibleKeyButtonView extends KeyButtonView {
                     // Let's hope we don't catch one!
                 }
                 return true;
-        	} else {  // we must have a custom uri
-       		 	try {
-       		 		Intent intent = Intent.parseUri(mLongpress, 0);
-       		 		intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-       		 		getContext().startActivity(intent);
-       		 	} catch (URISyntaxException e) {
-       		 		Log.e(TAG, "URISyntaxException: [" + mLongpress + "]");
-       		 	} catch (ActivityNotFoundException e){
-       		 		Log.e(TAG, "ActivityNotFound: [" + mLongpress + "]");
-       		 	}
-       		 	return true;
-        	}        	           
+                } else {  // we must have a custom uri
+       	                try {
+       	                        Intent intent = Intent.parseUri(mLongpress, 0);
+       	                        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+       	                        getContext().startActivity(intent);
+       	                } catch (URISyntaxException e) {
+       	                        Log.e(TAG, "URISyntaxException: [" + mLongpress + "]");
+       	                } catch (ActivityNotFoundException e){
+       	                        Log.e(TAG, "ActivityNotFound: [" + mLongpress + "]");
+       	                }
+       	                return true;
+                }
         }
     };
